requirements.txt
python-dotenv
requests
web3-ethereum-defi[data]
flask
